dto para form de criacao de usuario builder para salva

+++ src/main/java/br/edu/ifsp/spo/bike_integration/repository/SenhaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.ifsp.spo.bike_integration.model.Senha;
 
-public interface SenhaRepository extends JpaRepository<Senha, Long>{
+public interface SenhaRepository extends JpaRepository<Senha, Integer>{
 	
 }
