+++ src/main/java/br/edu/ifsp/spo/bike_integration/repository/WebServiceTokenRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.edu.ifsp.spo.bike_integration.model.WebServiceToken;
 
-public interface WebServiceTokenRepository extends JpaRepository<WebServiceToken, Long> {
+public interface WebServiceTokenRepository extends JpaRepository<WebServiceToken, Integer> {
     Optional<WebServiceToken> findByToken(String token);
 }
